Basic implementation of Game Inspector
Please note that a lot of this code is unfinished and no way near final.
- Implemented a Game Inspector window with two panels:
-- Hierarchy. Similar to the hierarchy in Unity, it shows a tree of all
gameobjects in the scene. If you click one it will show up in the
Inspector. It has a search field which will filter gameobjects in the
list but it's not functional yet.
-- Inspector. Allows you to inspect gameobjects. Lets you change name of
them and view their components. You can expand components to view their
public fields and properties and it currently lets you modify variables
of the type String and Vector3, more are planned.